Move test utilities into src/xcp_d/_testing.py
Extract mock_config, get_nodes, chdir, modified_environ,
get_cpu_count, and update_resources from test/tests.py and
test/utils.py into a new src/xcp_d/_testing.py module.

This eliminates imports from the test/ package in both test files
and workflow docstrings, and makes these utilities available as
part of the installed package under xcp_d._testing. Delete
test/tests.py now that all its content has moved.
