+++ src/xcp_d/workflows/anatomical/plotting.py
@@ -42,7 +42,7 @@ def init_brainsprite_figures_wf(
             :graph2use: orig
             :simple_form: yes
 
-            from test.tests import mock_config
+            from xcp_d._testing import mock_config
             from xcp_d import config
             from xcp_d.workflows.anatomical.plotting import init_brainsprite_figures_wf
 
@@ -264,7 +264,7 @@ def init_itk_warp_gifti_surface_wf(name='itk_warp_gifti_surface_wf'):
             :graph2use: orig
             :simple_form: yes
 
-            from test.tests import mock_config
+            from xcp_d._testing import mock_config
             from xcp_d import config
             from xcp_d.workflows.anatomical.plotting import init_itk_warp_gifti_surface_wf
 
@@ -356,7 +356,7 @@ def init_execsummary_anatomical_plots_wf(
             :graph2use: orig
             :simple_form: yes
 
-            from test.tests import mock_config
+            from xcp_d._testing import mock_config
             from xcp_d import config
             from xcp_d.workflows.anatomical.plotting import init_execsummary_anatomical_plots_wf
