+++ src/xcp_d/workflows/base.py
@@ -63,7 +63,7 @@ def init_xcpd_wf():
             :graph2use: orig
             :simple_form: yes
 
-            from test.tests import mock_config
+            from xcp_d._testing import mock_config
             from xcp_d import config
             from xcp_d.workflows.base import init_xcpd_wf
 
@@ -111,7 +111,7 @@ def init_single_subject_wf(subject_id: str, anat_session: str, func_sessions: li
             :graph2use: orig
             :simple_form: yes
 
-            from test.tests import mock_config
+            from xcp_d._testing import mock_config
             from xcp_d import config
             from xcp_d.workflows.base import init_single_subject_wf
