+++ src/xcp_d/workflows/bold/cifti.py
@@ -48,7 +48,7 @@ def init_postprocess_cifti_wf(
 
             import os
 
-            from test.tests import mock_config
+            from xcp_d._testing import mock_config
             from xcp_d import config
             from xcp_d.utils.bids import collect_data, collect_run_data
             from xcp_d.workflows.bold.cifti import init_postprocess_cifti_wf
