+++ src/xcp_d/workflows/bold/concatenation.py
@@ -31,7 +31,7 @@ def init_concatenate_data_wf(TR, head_radius, exact_scans, mem_gb, name='concate
             :graph2use: orig
             :simple_form: yes
 
-            from test.tests import mock_config
+            from xcp_d._testing import mock_config
             from xcp_d import config
             from xcp_d.workflows.bold.concatenation import init_concatenate_data_wf
