+++ src/xcp_d/workflows/parcellation.py
@@ -28,7 +28,7 @@ def init_load_atlases_wf(name='load_atlases_wf'):
             :graph2use: orig
             :simple_form: yes
 
-            from test.tests import mock_config
+            from xcp_d._testing import mock_config
             from xcp_d import config
             from xcp_d.data import load as load_data
             from xcp_d.workflows.parcellation import init_load_atlases_wf
@@ -267,7 +267,7 @@ def init_parcellate_cifti_wf(
             :graph2use: orig
             :simple_form: yes
 
-            from test.tests import mock_config
+            from xcp_d._testing import mock_config
             from xcp_d import config
             from xcp_d.workflows.parcellation import init_parcellate_cifti_wf
