+++ test/test_cli.py
@@ -11,14 +11,13 @@ import pytest
 from nipype import logging
 
 from test.utils import (
-    chdir,
     check_affines,
     check_generated_files,
     download_test_data,
     get_test_data_path,
     list_files,
-    update_resources,
 )
+from xcp_d._testing import chdir, update_resources
 from xcp_d.cli import combineqc, run
 from xcp_d.cli.parser import parse_args
 from xcp_d.cli.workflow import build_boilerplate, build_workflow
