+++ test/test_utils_execsummary.py
@@ -4,7 +4,7 @@ import os
 
 import matplotlib.pyplot as plt
 
-from test.utils import chdir
+from xcp_d._testing import chdir
 from xcp_d.data import load as load_data
 from xcp_d.utils import execsummary
