+++ test/test_workflows_metrics.py
@@ -6,9 +6,9 @@ import shutil
 import nibabel as nb
 import numpy as np
 
-from test.tests import mock_config
 from test.utils import get_nodes
 from xcp_d import config
+from xcp_d._testing import mock_config
 from xcp_d.utils.bids import _get_tr
 from xcp_d.utils.utils import _create_mem_gb
 from xcp_d.utils.write_save import read_ndata, write_ndata
